pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-github-review-server"
version = "0.1.0"
description = "MCP server for managing GitHub pull request reviews"
authors = [
  { name = "OpenAI MCP", email = "ops@example.com" }
]
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "httpx>=0.26",
  "anyio>=3.7",
  "fastapi>=0.111",
  "uvicorn>=0.27",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "structlog>=24.1",
  "python-json-logger>=2.0",
  "cachetools>=5.3",
  "tenacity>=8.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "respx>=0.20",
  "coverage[toml]>=7.4",
  "ruff>=0.4",
  "mypy>=1.8",
  "types-PyYAML",
  "types-cachetools"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["mcp_server"]
branch = true

[tool.coverage.report]
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "N"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true

